% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipHeatmap.R
\name{ChipHeatmap}
\alias{ChipHeatmap}
\title{Create a heatmap of chip data}
\usage{
ChipHeatmap(
  norm.mat = NULL,
  sample.label = "sample name",
  label.rect.gp = gpar(fill = "grey90"),
  label.gp = gpar(fontface = "bold"),
  row.split = NULL,
  row.order = NULL,
  ht.col = c("white", "red"),
  ht.col.legend.nbreaks = 4,
  ht.xaxis.rot = 0,
  col.range = NULL,
  heatmap_rank_method = c("weighting", "sum", "mean", "median"),
  draw.legend = TRUE,
  legend.pos = c("right", "bottom"),
  cluster.gap = 0.01,
  keep.cluster.panel.same = FALSE,
  quantile.threshold = 1,
  draw.anno.left = TRUE,
  draw.anno.left.shape = c("tri", "rect"),
  draw.anno.left.diretion = c("down", "up"),
  anno.left.col = NULL,
  draw.profile = FALSE,
  add.profile.rect = TRUE,
  profile.line.col = NULL,
  profile.line.size = 1,
  profile.y.range = NULL,
  profile.yaxis = TRUE,
  draw.line.legend = TRUE,
  profile.legend.size = c(0.25, 0.5),
  profile.legend.fontsize = 8,
  interpolate = FALSE,
  HeatmapLayout.params = list(),
  panel.col.space = 2,
  panel.row.space = 2
)
}
\arguments{
\item{norm.mat}{A numeric matrix containing normalized chip data which from
normalizeToMatrix/getTagMatrix/parseDeeptools.}

\item{sample.label}{A character string specifying the label for the sample.
Default is "sample name".}

\item{label.rect.gp}{A \code{gpar} object specifying the graphic parameters
for the label rectangle. Default is \code{gpar(fill = "grey90")}.}

\item{label.gp}{A \code{gpar} object specifying the graphic parameters for
the label. Default is \code{gpar(fontface = "bold")}.}

\item{row.split}{A numeric vector specifying the row split index.
Default is \code{NULL}.}

\item{row.order}{A numeric vector specifying the row order index.
Default is \code{NULL}.}

\item{ht.col}{A character vector specifying the colors for heatmap.
Default is \code{c("white","red")}.}

\item{ht.col.legend.nbreaks}{The number of colorbar legend breaks. Default is 4.}

\item{ht.xaxis.rot}{A numeric value specifying the rotation angle of
the x-axis labels. Default is 0.}

\item{col.range}{A numeric vector specifying the range of colors for
the heatmap. Default is \code{NULL}.}

\item{heatmap_rank_method}{A character string specifying the method for
ranking the heatmap. The available options are "weighting", "sum",
"mean", and "median". Default is "weighting". Note "weighting" can not
be used for data which is from "getTagMatrix" function.}

\item{draw.legend}{A logical value indicating whether to draw the legend.
Default is \code{TRUE}.}

\item{legend.pos}{A character vector specifying the position of the legend.
Default is \code{c("right","bottom")}.}

\item{cluster.gap}{A numeric value specifying the gap between clusters.
Default is 0.01.}

\item{keep.cluster.panel.same}{A logical value indicating whether to keep
the cluster panel the same. Default is \code{FALSE}.}

\item{quantile.threshold}{A numeric value specifying the quantile threshold.
Default is 1.}

\item{draw.anno.left}{A logical value indicating whether to draw the left
annotation. Default is \code{TRUE}.}

\item{draw.anno.left.shape}{A character vector specifying the shape of the
left annotation. Default is \code{c("tri","rect")}.}

\item{draw.anno.left.diretion}{The arrow of heatmap left annotation
bar direction "down"(default) or "up".}

\item{anno.left.col}{A character vector specifying the color of the left
annotation. Default is \code{NULL}.}

\item{draw.profile}{A logical value indicating whether to draw the profile.
Default is \code{FALSE}.}

\item{add.profile.rect}{Whether add rect border for the profile plot.
Default is TRUE.}

\item{profile.line.col}{A character vector specifying the color of the
profile line. Default is \code{NULL}.}

\item{profile.line.size}{A numeric value specifying the size of the profile
line. Default is 1.}

\item{profile.y.range}{A numeric vector specifying the y-range of the
profile. Default is \code{NULL}.}

\item{profile.yaxis}{A logical value indicating whether to draw the y-axis
of the profile. Default is \code{TRUE}.}

\item{draw.line.legend}{A logical value indicating whether to draw the line
legend. Default is \code{TRUE}.}

\item{profile.legend.size}{A numeric vector specifying the size of the line
legend. Default is \code{c(0.25,0.5)}.}

\item{profile.legend.fontsize}{A numeric value specifying the font size of
the line legend. Default is 8.}

\item{interpolate}{A logical value indicating whether to interpolate the
heatmap. Default is \code{FALSE}.}

\item{HeatmapLayout.params}{A list specifying the parameters for the heatmap
layout. Default is \code{list()}.}

\item{panel.col.space}{A numeric value specifying the space between panels
in the column direction. Default is 2.}

\item{panel.row.space}{A numeric value specifying the space between panels
in the row direction. Default is 2.}
}
\value{
A gtable object representing the heatmap.
}
\description{
This function creates a single heatmap of chip data using grid.
}
\author{
Jun Zhang
}
