% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{draw.anno.fun}
\alias{draw.anno.fun}
\title{Draw sample annnotations}
\usage{
draw.anno.fun(
  group.list = NULL,
  row.pos = 1,
  group.anno.rect.fill = NULL,
  group.anno.rect.col = NULL,
  group.anno.line.col = NULL,
  group.anno.line.lwd = 2,
  group.anno.text.face = "bold",
  group.anno.text.size = 12,
  sample.anno.type = c("rect", "line")
)
}
\arguments{
\item{group.list}{group.list}

\item{row.pos}{row.pos}

\item{group.anno.rect.fill}{group.anno.rect.fill}

\item{group.anno.rect.col}{group.anno.rect.col}

\item{group.anno.line.col}{group.anno.line.col}

\item{group.anno.line.lwd}{group.anno.line.lwd}

\item{group.anno.text.face}{group.anno.text.face}

\item{group.anno.text.size}{group.anno.text.size}

\item{sample.anno.type}{sample.anno.type}
}
\value{
grid object
}
\description{
Draw sample annnotations
}
