% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{get.col.average}
\alias{get.col.average}
\title{Calculate mean/median for columns of different rowsplit clusters}
\usage{
get.col.average(
  norm.mat = NULL,
  row.split = NULL,
  fun = c("mean", "median"),
  quantile.threshold = 1
)
}
\arguments{
\item{norm.mat}{norm.mat}

\item{row.split}{row.split}

\item{fun}{fun}

\item{quantile.threshold}{quantile.threshold}
}
\value{
range of value
}
\description{
Calculate mean/median for columns of different rowsplit clusters
}
