% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{get.group.bound}
\alias{get.group.bound}
\title{Getting panel boundary from group list}
\usage{
get.group.bound(group = NULL)
}
\arguments{
\item{group}{goup list}
}
\value{
list of boundary
}
\description{
Getting panel boundary from group list
}
