% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{get.matlist.range}
\alias{get.matlist.range}
\title{Getting whole range from list of matrice}
\usage{
get.matlist.range(
  mat.list = NULL,
  type = c("raw", "ave"),
  quantile.threshold = 1,
  row.split = NULL
)
}
\arguments{
\item{mat.list}{mat.list}

\item{type}{the type to calculate range, "raw" for heatmap, "ave" for profile.}

\item{quantile.threshold}{quantile.threshold}

\item{row.split}{row.split}
}
\value{
range of value
}
\description{
Getting whole range from list of matrice
}
