% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{get.row.order}
\alias{get.row.order}
\title{Getting ranked row orders from a matrix}
\usage{
get.row.order(
  norm.mat = NULL,
  row.split = NULL,
  heatmap_rank_method = NULL,
  quantile.threshold = 1
)
}
\arguments{
\item{norm.mat}{normalized matrix with attributes}

\item{row.split}{row.split}

\item{heatmap_rank_method}{heatmap_rank_method}

\item{quantile.threshold}{quantile.threshold}
}
\value{
list orders
}
\description{
Getting ranked row orders from a matrix
}
