% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{getAttrs}
\alias{getAttrs}
\title{Geeting attributes from matrix}
\usage{
getAttrs(mat = NULL)
}
\arguments{
\item{mat}{normalizedMatrix/getTagMatrix/parseDeeptools object}
}
\value{
axis_name breaks vline.x text.x.hjust
}
\description{
Geeting attributes from matrix
}
