% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobFuncs.R
\name{grid.colorkey}
\alias{grid.colorkey}
\title{Create a Color Key for Grid Graphics}
\usage{
grid.colorkey(
  x = NULL,
  color = NULL,
  color.n = 100,
  value.limits = NULL,
  ticks.side = c("left", "right", "top", "bottom"),
  pos = c("v", "h"),
  nbreaks = 5,
  draw = TRUE,
  xaxis2.params = list(),
  yaxis2.params = list(),
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{The values for which colors will be displayed. Default is NULL.}

\item{color}{The colors to be used in the color key. Default is NULL.}

\item{color.n}{The number of colors to generate. Default is 100.}

\item{value.limits}{The limits of the values. Default is NULL.}

\item{ticks.side}{The side on which to display tick marks. Default is "left".}

\item{pos}{The position of the color key ('v' for vertical, 'h' for horizontal).
Default is "v".}

\item{nbreaks}{The number of breaks for the color key. Default is 5.}

\item{draw}{Logical. Should the color key be drawn immediately? Default is TRUE.}

\item{xaxis2.params}{Parameters for the x-axis. Default is an empty list.}

\item{yaxis2.params}{Parameters for the y-axis. Default is an empty list.}

\item{name}{The name of the color key. Default is NULL.}

\item{gp}{A graphical parameter object. Default is NULL.}

\item{vp}{A viewport object. Default is NULL.}
}
\value{
If \code{draw} is \code{TRUE}, the color key is drawn. Otherwise,
a grob is returned.
}
\description{
This function creates a color key for use in grid graphics.
}
\seealso{
\code{\link[grid]{grid.draw}}, \code{\link[grid]{gTree}}
}
