% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapLayout.R
\name{HeatmapLayout}
\alias{HeatmapLayout}
\title{Heatmap Layout Function}
\usage{
HeatmapLayout(
  mat = NULL,
  heatmap.size = c(0.7, 0.5),
  heatmap.xaxis.h = 0.1,
  heatmap.blegend.h = 0.035,
  heatmap.rlegend.w = 0.035,
  heatmap.profile.h = 0.2,
  heatmap.label.h = 0.075,
  heatmap.lanno.w = 0.05,
  heatmap.lanno.label.w = 0.1,
  draw.profile = TRUE,
  legend.pos = c("right", "bottom"),
  plot.size = c(1, 1),
  newpage = TRUE,
  draw.layout = FALSE
)
}
\arguments{
\item{mat}{A matrix for the heatmap data which from
normalizeToMatrix/getTagMatrix/parseDeeptools.}

\item{heatmap.size}{A numeric vector specifying the size of the heatmap in
relative units (default: c(0.7, 0.5)).}

\item{heatmap.xaxis.h}{The height of the heatmap x-axis labels in relative
units (default: 0.1).}

\item{heatmap.blegend.h}{The height of the heatmap bottom legend in relative
units (default: 0.035).}

\item{heatmap.rlegend.w}{The width of the heatmap right legend in relative
units (default: 0.035).}

\item{heatmap.profile.h}{The height of the heatmap profile area in relative
units (default: 0.2).}

\item{heatmap.label.h}{The height of the heatmap label area in relative
units (default: 0.075).}

\item{heatmap.lanno.w}{The width of the heatmap left annotation area in
relative units (default: 0.05).}

\item{heatmap.lanno.label.w}{The width of the heatmap left annotation labels
in relative units (default: 0.1).}

\item{draw.profile}{Logical indicating whether to draw the heatmap profile
area (default: TRUE).}

\item{legend.pos}{The position of the heatmap legend ("right" or "bottom",
default: "right").}

\item{plot.size}{A numeric vector specifying the size of the overall plot
(default: c(1, 1)).}

\item{newpage}{Logical indicating whether to start a new page for the plot
(default: TRUE).}

\item{draw.layout}{Logical indicating whether to draw the heatmap layout
(default: FALSE).}
}
\value{
A gtable object representing the heatmap layout.
}
\description{
This function creates a layout for a heatmap plot with using "gtabel" R
package. And this is used in "ChipHeatmap" function.
}
\author{
Jun Zhang
}
