% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visFuncUtils.R
\name{match.col2val}
\alias{match.col2val}
\title{Mapping a numeric range to colors}
\usage{
match.col2val(cols = c("white", "red"), cols.n = 100, range.val = NULL)
}
\arguments{
\item{cols}{the colors to be mapped, default c("white","red").}

\item{cols.n}{the numbers color to be generated, default 100.}

\item{range.val}{the data range to be mapped.}
}
\value{
a data.frame
}
\description{
Mapping a numeric range to colors
}
