% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiHeatmap.R
\name{multiHeatmap}
\alias{multiHeatmap}
\title{MultiHeatmap}
\usage{
multiHeatmap(
  mat.list = NULL,
  row.split = NULL,
  sample.label = NULL,
  heatmap_rank_method = c("weighting", "sum", "mean", "median"),
  quantile.threshold = 0.99,
  add.panel.rect = FALSE,
  panel.rect.gp = gpar(),
  keep.row.order = TRUE,
  scale.range = FALSE,
  heatmap.col = NULL,
  keep.one.left.annobar = FALSE,
  keep.one.legend = FALSE,
  ChipHeatmap.params = list(),
  respect = FALSE,
  plot.size = c(0.9, 0.9),
  draw.profile = FALSE,
  keep.one.line.legend = FALSE,
  keep.one.profile.yaxis = FALSE,
  scale.y.range = FALSE,
  sample.group1 = NULL,
  sample.group2 = NULL,
  anno.panel.height = 0.1,
  draw.anno.fun.params.g1 = list(),
  draw.anno.fun.params.g2 = list()
)
}
\arguments{
\item{mat.list}{A list of matrices to be plotted which from
normalizeToMatrix/getTagMatrix/parseDeeptools.}

\item{row.split}{A vector specifying the groups to which each row belongs.
This vector should be of the same length as the number of rows in the matrices.
Default is NULL.}

\item{sample.label}{A vector of labels for the samples. This vector should
be of the same length as mat.list. Default is NULL.}

\item{heatmap_rank_method}{A character string specifying the method to rank
the rows in the heatmap. The available options are "weighting", "sum",
"mean", and "median". Default is "weighting". Note "weighting" can not
be used for data which is from "getTagMatrix" function.}

\item{quantile.threshold}{A numeric value specifying the quantile threshold
for determining the range of the color scale. Default is 0.99.}

\item{add.panel.rect}{Logical value indicating whether to add a rectangle
around the entire heatmap plot. Default is FALSE.}

\item{panel.rect.gp}{An object of class "gpar" specifying graphical
parameters for the panel rectangle. Default is gpar().}

\item{keep.row.order}{Logical value indicating whether to keep the row order
of the first plot for the rest of the plots. Default is TRUE.}

\item{scale.range}{Logical value indicating whether to scale the color range
of each heatmap based on the quantile threshold. Default is FALSE.}

\item{heatmap.col}{A list of color schemes for each heatmap. If NULL, the
default color scheme is used. Default is NULL.}

\item{keep.one.left.annobar}{Logical value indicating whether to keep only
one left annotation panel for all heatmaps. Default is FALSE.}

\item{keep.one.legend}{Logical value indicating whether to keep only one
legend panel for all heatmaps. Default is FALSE.}

\item{ChipHeatmap.params}{A list of additional parameters to be passed to
the ChipHeatmap function. Default is list().}

\item{respect}{Logical value indicating whether to respect the layout
parameters specified by widths and heights when using multiple heatmaps.
Default is FALSE.}

\item{plot.size}{A numeric vector specifying the size of the overall plot.
The first element represents the width and the second element represents
the height. Default is c(0.9,0.9).}

\item{draw.profile}{Logical value indicating whether to draw a profile plot
below each heatmap. Default is FALSE.}

\item{keep.one.line.legend}{Logical value indicating whether to keep only
one line legend panel for all profile plots. Default is FALSE.}

\item{keep.one.profile.yaxis}{Logical value indicating whether to keep only
one y-axis label for all profile plots. Default is FALSE.}

\item{scale.y.range}{Logical value indicating whether to scale the y-axis
range of each profile plot based on the quantile threshold. Default is FALSE.}

\item{sample.group1}{A vector specifying the grouping of samples for the
first annotation panel. Default is NULL.}

\item{sample.group2}{A vector specifying the grouping of samples for the
second annotation panel. Default is NULL.}

\item{anno.panel.height}{Numeric value specifying the relative height of
the annotation panel. Default is 0.1.}

\item{draw.anno.fun.params.g1}{A list of additional parameters to be passed
to the annotation function for the first group of samples.
Default is list(), details see draw.anno.fun().}

\item{draw.anno.fun.params.g2}{A list of additional parameters to be passed
to the annotation function for the second group of samples.
Default is list(), details see draw.anno.fun().}
}
\value{
None
}
\description{
This function creates a multi-panel heatmap plot based on multiple matrices
using ChipHeatmap function.
}
\author{
Jun Zhang
}
