% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDeeptools.R
\name{parseDeeptools}
\alias{parseDeeptools}
\title{Parse deeptools output}
\usage{
parseDeeptools(deeptools_output = NULL)
}
\arguments{
\item{deeptools_output}{Path to the deeptools output file.}
}
\value{
A list of matrices with additional attributes.
}
\description{
This function parses the deeptools output and returns a list of matrices.
Each matrix represents the data for a specific sample and contains additional
attributes like upstream, downstream, extend, and target indices.
}
\details{
The output of deeptools in gzipped text format is read using \code{data.table::fread()}.
The function extracts meta information from the first column of the data and
processes it to get relevant values for upstream, downstream, body, bin size,
group labels, group boundaries, sample labels, and sample boundaries.
Then, it loops through the sample labels and creates a matrix for each sample,
filling any missing values with zeros. Lastly, it assigns additional attributes
to each matrix based on the extracted values and returns a list of matrices.
}
\author{
Jun Zhang
}
