% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriveData.R
\name{retriveData}
\alias{retriveData}
\title{Retrieve Data}
\usage{
retriveData(
  mat.list = NULL,
  row.split = NULL,
  sample.names = NULL,
  rm.extreme.value = FALSE,
  group.sample = NULL,
  group2.sample = NULL,
  aggregate.fun = c("mean", "median"),
  ChIPseeker = FALSE,
  heatmap_rank_method = c("sum", "mean", "median", "weighting")
)
}
\arguments{
\item{mat.list}{A list of matrices containing the data which output from
normalizeToMatrix/getTagMatrix/parseDeeptools.}

\item{row.split}{A character vector specifying the group of the peaks. This will use
the raw peak sets when data is from "deeptools" output.}

\item{sample.names}{A character vector specifying the names of the samples.}

\item{rm.extreme.value}{Logical value indicating whether to remove extreme binding
signal values. Default FALSE.}

\item{group.sample}{A character vector specifying the group of each sample.}

\item{group2.sample}{A character vector specifying the second group of each sample.}

\item{aggregate.fun}{A character vector specifying the aggregate function(mean/meidan) to
use for profile data. Default "mean".}

\item{ChIPseeker}{Logical value indicating whether "ChIPseeker" output data is used.}

\item{heatmap_rank_method}{The method for heatmap rows rank. "sum","mean","median" and
"weighting" can be used. "weighting" method is from EnrichedHeatmap::enriched_score function
and this is applied when data is from EnrichedHeatmap::normalizeToMatrix. Default "sum".}
}
\value{
A list containing the profile and heatmap data.
}
\description{
This function retrieves profile and heatmap data (EnrichedHeatmap::normalizeToMatrix,
ChIPseeker::getTagMatrix and deeptools' computeMatrix output) from a list of matrices.
Some codes origin from R package EnrichedHeatmap/ChIPseeker with some modifications.
Thanks for their contributions!
}
\author{
Jun Zhang
}
